Added vagrant v2 vm config monkey patch
This patch essentially reverts the shared folder behavior of vagrant 1.2 to 1.1. This is needed so that Windows guests can properly map a Windows share, as Windows shares cannot contain slashes in the name. The share id is a short name which is a valid share, so we'll use that just like vagrant 1.1 did.

See vagrant issue 1742 for more info. https://github.com/mitchellh/vagrant/issues/1742