Fixed issue 104
The WinRM.host config value now overrides the host value returned from the machine's ssh_info(). To enable this the default value for the winrm.host was changed to null since we can lean on the ssh_info[:host] default value and we need to know if the user explicitly set a value in the Vagranfile config.

Added unit tests to cover the winrmfinder host method.